[FlexNG] Use previous constraint space for layout roots.
Devtools had an issue where element were size at 0px height.
This was caused by the relayout boundary logic not working correctly.

In Devtools these elements had layout and size containment triggering
ObjectIsRelayoutBoundary() to return true.

The call to NGConstraintSpace::CreateFromLayoutObject for this flex-item
was incorrect as it wasn't using the stretched size (as we don't
set HasOverrideLogicalWidth(), etc).

This refactors the various callers of
NGConstraintSpace::CreateFromLayoutObject and create a new method
UpdateInFlowBlockLayout which contains the previously (duplicated)
logic.

This new method will try and use the previous constraint space (if
it is a layout root), instead of creating it.

Bug: 845235
Change-Id: I52ddb9d58ae010305b980657bb434a30d5b7fbbb