Fire "load" event at <style> when textContent changes
Prior to this CL, the load event was not being fired when the
textContent of a <style> element was changed, contrary to the
spec ([1]). Now, the load event is fired in this case.

Note that when textContent is changed, ContainerNode::RemoveChildren
first removes the existing stylesheet contents, which in turn
creates a new (empty) stylesheet at [2], and that action would
normally have also fired a load event at the temporarily-empty
style element. The |removing_all_children_| flag keeps track of
this situation and prevents the duplicate event.

[1] https://html.spec.whatwg.org/multipage/semantics.html#update-a-style-block
[2] https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/renderer/core/css/style_element.cc;l=97;drc=bbbbfe1fc2ec68a51ba159891b54d9e19c34972f

Fixed: 674518

Change-Id: Idb0ddb9f0ca2478a8a35d06817928503e74fa3fb
Bug: 674518